+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/web_link.php
@@ -1,5 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'web_link' => ['enabled' => true],
-]);