+++ src/Symfony/Component/WebLink/Tests/DependencyInjection/Fixtures/web_link.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('web_link', null);